+++ SharePoint/SharePoint-ADFS/CHANGELOG.md
@@ -11,6 +11,14 @@
 * Updated resource SPTrustedIdentityTokenIssuer in SharePointDSC 2.5 to support parameter UseWReplyParameter
 * Updated SharePointDSC from 2.2 to 2.5 (with custom fix on resource SPUserProfileServiceApp)
 * Updated xDnsServer from 1.10 to 1.11
+* Updated xActiveDirectory from 2.18 to 2.21
+* Updated xPSDesiredStateConfiguration from 8.2 to 8.4
+* Updated ActiveDirectoryCSDsc from 2.0 to 3.0
+* Updated ComputerManagementDsc from 5.0 to 5.2
+* Updated SqlServerDsc from 11.2 to 12.0
+* Updated CertificateDsc from 4.0 to 4.2
+* Updated xWebAdministration from 1.20 to 2.2
+* Updated from xNetworking 5.7 to NetworkingDsc 6.1
 
 ## June 2018 update